python/requirements.txt
# Core dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9

# OpenAI integration
openai>=1.55.0

# Document processing
python-docx>=1.1.2
docx2txt>=0.8
openpyxl>=3.1.5

# Configuration and validation
python-dotenv>=1.0.1
pydantic>=2.10.0

# File handling
aiofiles>=24.1.0